+++ packages/foam-vscode/src/features/index.ts
@@ -9,7 +9,7 @@ import createFromTemplate from './create-from-template';
 import openRandomNote from './open-random-note';
 import orphans from './orphans';
 import placeholders from './placeholders';
-import filteredResources from './filtered-resources';
+import utilityCommands from './utility-commands';
 import { FoamFeature } from '../types';
 
 export const features: FoamFeature[] = [
@@ -24,5 +24,5 @@ export const features: FoamFeature[] = [
   createFromTemplate,
   orphans,
   placeholders,
-  filteredResources,
+  utilityCommands,
 ];
