+++ packages/foam-vscode/src/features/orphans.ts
@@ -1,8 +1,8 @@
 import * as vscode from 'vscode';
 import { Foam, FoamWorkspace, isNote, Resource } from 'foam-core';
-import { FilteredResourcesConfigGroupBy, getOrphansConfig } from '../settings';
+import { getOrphansConfig } from '../settings';
 import { FoamFeature } from '../types';
-import { FilteredResourcesProvider } from './filtered-resources';
+import { GroupedResourcesTreeDataProvider } from '../utils/grouped-resources-tree-data-provider';
 
 const feature: FoamFeature = {
   activate: async (
@@ -10,33 +10,24 @@ const feature: FoamFeature = {
     foamPromise: Promise<Foam>
   ) => {
     const foam = await foamPromise;
+
     const workspacesURIs = vscode.workspace.workspaceFolders.map(
       dir => dir.uri
     );
-    const provider = new FilteredResourcesProvider(
+
+    const provider = new GroupedResourcesTreeDataProvider(
       foam.workspace,
       foam.services.dataStore,
       'orphans',
       'orphan',
       (resource: Resource) => isOrphan(foam.workspace, resource),
-      {
-        ...getOrphansConfig(),
-        workspacesURIs,
-      }
+      getOrphansConfig(),
+      workspacesURIs
     );
 
     context.subscriptions.push(
       vscode.window.registerTreeDataProvider('foam-vscode.orphans', provider),
-      vscode.commands.registerCommand(
-        'foam-vscode.group-orphans-by-folder',
-        () => provider.setGroupBy(FilteredResourcesConfigGroupBy.Folder)
-      ),
-      vscode.commands.registerCommand('foam-vscode.group-orphans-off', () =>
-        provider.setGroupBy(FilteredResourcesConfigGroupBy.Off)
-      ),
-      foam.workspace.onDidAdd(() => provider.refresh()),
-      foam.workspace.onDidUpdate(() => provider.refresh()),
-      foam.workspace.onDidDelete(() => provider.refresh())
+      ...provider.commands
     );
   },
 };
